add some routes for different maps
