+++ pages/maps/sky-map.tsx
@@ -1,0 +1,10 @@
+import * as React from 'react'
+import LayerMap from '@/components/LayerMap'
+
+function SkyMapPage() {
+  return (
+    <LayerMap />
+  )
+}
+
+export default SkyMapPage
